# frontend/docs/api/试题信息管理_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "🎓 智能考试系统API文档",
    "description": "📚 智能考试系统后端接口文档，提供完整的RESTful API服务\n\n✨ 主要功能模块：\n• 🧠 题目管理：支持选择题、判断题、简答题的增删改查\n• 📝 试卷管理：手动组卷和AI智能组卷\n• 🎨 轮播图管理：首页轮播图的图片上传和管理\n• 📊 考试记录：考试结果统计和分析\n• 🔔 公告管理：系统公告的发布和管理",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "分类管理"
        ],
        "summary": "获取分类列表",
        "description": "获取所有题目分类列表，包含每个分类下的题目数量统计",
        "operationId": "getCategories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListCategory"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "分类管理"
        ],
        "summary": "更新分类信息",
        "description": "修改分类的名称、描述、排序等信息",
        "operationId": "updateCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "分类管理"
        ],
        "summary": "添加新分类",
        "description": "创建新的题目分类，支持设置父分类实现层级结构",
        "operationId": "addCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions/category/{categoryId}": {
      "get": {
        "tags": [
          "题目管理"
        ],
        "summary": "按分类查询题目",
        "description": "获取指定分类下的所有题目列表",
        "operationId": "getQuestionsByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "分类ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListQuestion"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/categories/{id}": {
      "delete": {
        "tags": [
          "分类管理"
        ],
        "summary": "删除分类",
        "description": "删除指定的题目分类，注意：删除前需确保分类下没有题目",
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "分类ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "主键",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "修改时间",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "分类名称",
            "example": "Java基础"
          },
          "parentId": {
            "type": "integer",
            "description": "父分类ID，顶级分类为0",
            "format": "int64",
            "example": 0
          },
          "sort": {
            "type": "integer",
            "description": "排序序号，数字越小越靠前",
            "format": "int32",
            "example": 1
          },
          "children": {
            "type": "array",
            "description": "子分类列表，用于构建分类树结构",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "count": {
            "type": "integer",
            "description": "该分类下的题目数量",
            "format": "int64",
            "example": 25
          }
        },
        "description": "题目分类信息"
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "主键",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "修改时间",
            "format": "date-time"
          },
          "title": {
            "type": "string",
            "description": "题目标题内容",
            "example": "以下关于Java面向对象编程的说法正确的是？"
          },
          "type": {
            "type": "string",
            "description": "题目类型",
            "example": "CHOICE",
            "enum": [
              "CHOICE",
              "JUDGE",
              "TEXT"
            ]
          },
          "multi": {
            "type": "boolean",
            "description": "是否为多选题，仅选择题有效",
            "example": false
          },
          "categoryId": {
            "type": "integer",
            "description": "题目分类ID",
            "format": "int64",
            "example": 1
          },
          "difficulty": {
            "type": "string",
            "description": "题目难度等级",
            "example": "MEDIUM",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "score": {
            "type": "integer",
            "description": "题目默认分值",
            "format": "int32",
            "example": 5
          },
          "paperScore": {
            "type": "number",
            "description": "在特定试卷中的分值",
            "example": 10
          },
          "analysis": {
            "type": "string",
            "description": "题目解析，详细的答案说明",
            "example": "Java是面向对象编程语言，支持封装、继承、多态三大特性..."
          },
          "choices": {
            "type": "array",
            "description": "选择题选项列表，包含A、B、C、D等选项",
            "items": {
              "$ref": "#/components/schemas/QuestionChoice"
            }
          },
          "answer": {
            "$ref": "#/components/schemas/QuestionAnswer"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "description": "题目信息"
      },
      "QuestionAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "主键",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "修改时间",
            "format": "date-time"
          },
          "questionId": {
            "type": "integer",
            "description": "关联的题目ID",
            "format": "int64",
            "example": 1
          },
          "answer": {
            "type": "string",
            "description": "标准答案内容",
            "example": "正确"
          },
          "keywords": {
            "type": "string",
            "description": "评分关键词，用于简答题AI评分",
            "example": "面向对象,封装,继承,多态"
          }
        },
        "description": "题目答案信息"
      },
      "QuestionChoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "主键",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "修改时间",
            "format": "date-time"
          },
          "questionId": {
            "type": "integer",
            "description": "关联的题目ID",
            "format": "int64",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "选项内容",
            "example": "Java是面向对象编程语言"
          },
          "isCorrect": {
            "type": "boolean",
            "description": "是否为正确答案",
            "example": true
          },
          "sort": {
            "type": "integer",
            "description": "选项排序序号",
            "format": "int32",
            "example": 1
          }
        },
        "description": "题目选项信息"
      },
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码",
            "format": "int32",
            "example": 200,
            "enum": [
              200,
              400,
              401,
              403,
              404,
              500
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息，成功时描述操作结果，失败时描述错误原因",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据，具体的业务数据内容"
          }
        },
        "description": "统一API响应结果"
      },
...
      "ResultListQuestion": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码",
            "format": "int32",
            "example": 200,
            "enum": [
              200,
              400,
              401,
              403,
              404,
              500
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息，成功时描述操作结果，失败时描述错误原因",
            "example": "操作成功"
          },
          "data": {
            "type": "array",
            "description": "响应数据，具体的业务数据内容",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        },
        "description": "统一API响应结果"
      },
...
      "ResultListCategory": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码",
            "format": "int32",
            "example": 200,
            "enum": [
              200,
              400,
              401,
              403,
              404,
              500
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息，成功时描述操作结果，失败时描述错误原因",
            "example": "操作成功"
          },
          "data": {
            "type": "array",
            "description": "响应数据，具体的业务数据内容",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        },
        "description": "统一API响应结果"
      }
    }
  }
}